Refactor training pipeline modules
