+++ botcopier/scripts/cluster_regimes.py
@@ -23,7 +23,7 @@ except Exception:  # pragma: no cover - optional
 
 # Reuse data loading and feature extraction from training script
 from botcopier.data.loading import _load_calendar, _load_logs
-from botcopier.features.technical import _extract_features
+from botcopier.features.engineering import _extract_features
 
 
 def cluster_features(
