+++ tests/test_candlestick_patterns.py
@@ -5,7 +5,7 @@ import pandas as pd
 
 import botcopier.features.engineering as fe
 from botcopier.features.engineering import FeatureConfig, clear_cache, configure_cache
-from botcopier.features.technical import _extract_features
+from botcopier.features.engineering import _extract_features
 from botcopier.features.engineering import _extract_features as extract_rows_features
